requirements.txt
faiss-cpu == 1.12.0
langchain == 0.3.27
langchain-openai == 0.3.30
langchain-community == 0.3.27
matplotlib == 3.10.5
numpy == 1.3.2
openpyxl == 3.1.5
pandas == 2.3.1
streamlit == 1.48.1
tiktoken == 0.11.0
groq == 0.31.0